Another attempt at getting this running in a way that is at least somewhat reasonable in some way.
It's still far from ideal as I need to replicate type information a lot and have two traits for the same thing, but at least this is pretty close to seeming like something that might work with Rust's types. The only remaining problem (as far as the compiler will tell me right now) is the default implementation of generate_data. Apparently Rust complains that:

  `[<Self as test::TestImpl<A>>::D]` does not have a constant size known at compile-time

I have no idea how to resolve this. Stuffing the vector into a Box doesn't help.