Change NEXMark generation to use floats for the timestamp calculation instead.
This is a deviation from the original implementation in Beam as of time. The reason we are diverging is that the original implementation does not support rates higher than 2'000'000 events per second. At such a rate, the timestep becomes zero, meaning all events get the same timestamp. In order to support more fine-grained stepping than that we switch to a float-based scheme and only round when we deal out the timestamp to an event.